Finished implementing @loserskater building idea
Some differences, though.

Whenever we ADD a new file to this mods folder, we must first untrack it in the /system folder.

To do this, we run the git rm _____ command, where the space is the filepath and name.

This will stop any built files from being added to git unnecessarily.

We also add to .gitignore as well, so stop this tracking.

Both the compile_apps files can be run on command to test these folders out, and are integrated into the build files.

HIGHLY unrecommended to use these folders for developing. Instead, we should move completed mods from our development folders for apktool to the mod folder when they are finished and tested

Thanks again to loserskater for this idea :D