Kernel: Added choice between googymax or stock
Modified the Settings apks to look for the STweaks package being installed (added a method to do this). If STweaks isn't installed, it removes the "Performance" tile from the settings dashboard.

If a user installs stock, then Performance won't show for them in settings. If they ever decided to install Alucard or GoogyMax3, then performance would show again. Neat!

This is in response to people asking for stock kernel options, as they claim it's better for the audio stuttering on Bluetooth.