Update LINUX.md
Added schedtool to the necessary package. I totally forgot about that one!

Also I put the clone command in the wrong place in the git commands. I'll keep the single branch part for now, because in the future there will be more branches and it would be best to only clone the one we need, especially since this branch is about 2GB already!

This closes #21 as well.